Split asset overview into several files
